backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ravi Riley <raviriley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hume = "^0.4.1"
flask = "^3.0.2"
openai = "^1.12.0"
numpy = "^1.26.4"
flask-socketio = "^5.3.6"
pyserial = "^3.5"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"
mypy = "^1.8.0"
ipykernel = "^6.29.2"

[tool.poetry.scripts]
lint = "scripts.scripts:lint"
l = "scripts.scripts:lint"
format = "scripts.scripts:format_code"
f = "scripts.scripts:format_code"
format-lint = "scripts.scripts:format_and_lint"
fl = "scripts.scripts:format_and_lint"
test = "scripts.scripts:test"
t = "scripts.scripts:test"
format-lint-test = "scripts.scripts:format_lint_test"
flt = "scripts.scripts:format_lint_test"
dev = "scripts.scripts:dev"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
